Rename functions to improve readability
