+++ scripts/ont_append_fc_to_sample_name.py
@@ -5,7 +5,7 @@ from argparse import ArgumentParser
 from genologics.lims import Lims
 from genologics.config import BASEURI, USERNAME, PASSWORD
 from genologics.entities import Process
-from generate_ont_samplesheet import get_minknow_sample_id
+from ont_generate_samplesheet import get_minknow_sample_id
 
 DESC = """ Script for EPP "Append ONT flow cell to sample name".
 Append flow cell information from previous step to sample names in the current step.
