+++ scripts/ont_send_fc_to_db.py
@@ -11,7 +11,7 @@ import os
 import couchdb
 import yaml
 import sys
-from generate_ont_samplesheet import get_minknow_sample_id
+from ont_generate_samplesheet import get_minknow_sample_id
 
 
 DESC = """ Script for EPP "Send ONT flowcell info to StatusDB".
